backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.8.0,<2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0,<0.20.0
asyncpg>=0.24.0,<0.25.0
supabase>=2.0.0,<3.0.0
httpx>=0.24.0,<0.25.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.1,<5.0.0
email-validator>=1.1.3,<2.0.0
openai>=1.0.0,<2.0.0
anthropic>=0.5.0,<0.6.0 